fix: use u64 instead of usize in codec

+++ src/codec/body/payload_decoder.rs
@@ -37,7 +37,7 @@ impl PayloadDecoder {
 
     /// create a fixed length `PayloadDecoder`
     #[allow(unused)]
-    pub fn fix_length(size: usize) -> Self {
+    pub fn fix_length(size: u64) -> Self {
         Self { kind: Kind::Length(LengthDecoder::new(size)) }
     }
