+++ src/connection/connection.rs
@@ -134,14 +134,10 @@ where
             let size_hint = body.size_hint();
             match size_hint.exact() {
                 Some(0) => PayloadSize::Empty,
-                // todo usize maybe overflow
-                Some(length) => PayloadSize::Length(length as usize),
+                Some(length) => PayloadSize::Length(length),
                 None => PayloadSize::Chunked,
             }
         };
-
-        // Future
-        // Stream  Iterator
 
         self.framed_write.send(Message::Header((ResponseHead::from_parts(header_parts, ()), payload_size))).await?;
